MediaProvider acts as a really base provider for Media
We move all implementation to the superclass, now the code is more clear.
