+++ base/src/main/java/butter/droid/base/providers/media/MediaProvider.java
@@ -17,19 +17,26 @@
 
 package butter.droid.base.providers.media;
 
+import android.accounts.NetworkErrorException;
 import android.os.Parcel;
 import android.os.Parcelable;
 import android.support.annotation.DrawableRes;
 import android.support.annotation.Nullable;
 
 import com.squareup.okhttp.Call;
+import com.squareup.okhttp.Request;
+import com.squareup.okhttp.Response;
 
+import java.io.IOException;
+import java.util.AbstractMap;
 import java.util.ArrayList;
 import java.util.List;
 
+import butter.droid.base.R;
 import butter.droid.base.providers.BaseProvider;
 import butter.droid.base.providers.media.models.Genre;
 import butter.droid.base.providers.media.models.Media;
+import timber.log.Timber;
 
 /**
  * MediaProvider.java
@@ -37,6 +44,7 @@ import butter.droid.base.providers.media.models.Media;
  * Base class for all media providers. Any media providers has to extend this class and use the callback defined here.
  */
 public abstract class MediaProvider extends BaseProvider implements Parcelable {
+
     public static final String MEDIA_CALL = "media_http_call";
     public static final int DEFAULT_NAVIGATION_INDEX = 1;
     @SuppressWarnings("unused")
@@ -63,6 +71,17 @@ public abstract class MediaProvider extends BaseProvider implements Parcelable {
             return null;
         }
     };
+    private String[] apiUrls = new String[0];
+    private String itemsPath = "";
+    private String itemDetailsPath = "";
+    private Integer currentApi = 0;
+
+    public MediaProvider(String[] apiUrls, String itemsPath, String itemDetailsPath, Integer currentApi) {
+        this.apiUrls = apiUrls;
+        this.itemsPath = itemsPath;
+        this.itemDetailsPath = itemDetailsPath;
+        this.currentApi = currentApi;
+    }
 
     /**
      * Get a list of Media items from the provider
@@ -77,18 +96,193 @@ public abstract class MediaProvider extends BaseProvider implements Parcelable {
     /**
      * Get a list of Media items from the provider
      *
-     * @param currentList Input the current list so it can be extended
-     * @param filters     Filters the provider can use to sort or search
-     * @param callback    MediaProvider callback
+     * @param existingList Input the current list so it can be extended
+     * @param filters      Filters the provider can use to sort or search
+     * @param callback     MediaProvider callback
      * @return Call
      */
-    public abstract Call getList(ArrayList<Media> currentList, Filters filters, Callback callback);
+    public Call getList(final ArrayList<Media> existingList, Filters filters, final Callback callback) {
+        final ArrayList<Media> currentList;
+        if (existingList == null) {
+            currentList = new ArrayList<>();
+        } else {
+            currentList = new ArrayList<>(existingList);
+        }
 
-    public abstract Call getDetail(ArrayList<Media> currentList, Integer index, Callback callback);
+        ArrayList<AbstractMap.SimpleEntry<String, String>> params = new ArrayList<>();
+        params.add(new AbstractMap.SimpleEntry<>("limit", "30"));
 
-    public abstract int getLoadingMessage();
+        if (filters == null) {
+            filters = new Filters();
+        }
 
-    public abstract List<NavInfo> getNavigation();
+        if (filters.keywords != null) {
+            params.add(new AbstractMap.SimpleEntry<>("keywords", filters.keywords));
+        }
+
+        if (filters.genre != null) {
+            params.add(new AbstractMap.SimpleEntry<>("genre", filters.genre));
+        }
+
+        if (filters.order == Filters.Order.ASC) {
+            params.add(new AbstractMap.SimpleEntry<>("order", "1"));
+        } else {
+            params.add(new AbstractMap.SimpleEntry<>("order", "-1"));
+        }
+
+        if (filters.langCode != null) {
+            params.add(new AbstractMap.SimpleEntry<>("lang", filters.langCode));
+        }
+
+        String sort;
+        switch (filters.sort) {
+            default:
+            case POPULARITY:
+                sort = "popularity";
+                break;
+            case YEAR:
+                sort = "year";
+                break;
+            case DATE:
+                sort = "last added";
+                break;
+            case RATING:
+                sort = "rating";
+                break;
+            case ALPHABET:
+                sort = "name";
+                break;
+            case TRENDING:
+                sort = "trending";
+                break;
+        }
+
+        params.add(new AbstractMap.SimpleEntry<>("sort", sort));
+
+        String url = apiUrls[currentApi] + itemsPath;
+        if (filters.page != null) {
+            url += filters.page;
+        } else {
+            url += "1";
+        }
+
+        Request.Builder requestBuilder = new Request.Builder();
+        String query = buildQuery(params);
+        url = url + "?" + query;
+        requestBuilder.url(url);
+        requestBuilder.tag(MEDIA_CALL);
+
+        Timber.d(this.getClass().getSimpleName(), "Making request to: " + url);
+
+        return fetchList(currentList, requestBuilder, filters, callback);
+    }
+
+    /**
+     * Fetch the list of movies from API
+     *
+     * @param currentList    Current shown list to be extended
+     * @param requestBuilder Request to be executed
+     * @param callback       Network callback
+     * @return Call
+     */
+    private Call fetchList(final ArrayList<Media> currentList, final Request.Builder requestBuilder, final Filters filters, final Callback callback) {
+        return enqueue(requestBuilder.build(), new com.squareup.okhttp.Callback() {
+            @Override
+            public void onFailure(Request request, IOException e) {
+                String url = requestBuilder.build().urlString();
+                if (currentApi >= apiUrls.length - 1) {
+                    callback.onFailure(e);
+                } else {
+                    if (url.contains(apiUrls[currentApi])) {
+                        url = url.replace(apiUrls[currentApi], apiUrls[currentApi + 1]);
+                        currentApi++;
+                    } else {
+                        url = url.replace(apiUrls[currentApi - 1], apiUrls[currentApi]);
+                    }
+                    requestBuilder.url(url);
+                    fetchList(currentList, requestBuilder, filters, callback);
+                }
+            }
+
+            @Override
+            public void onResponse(Response response) throws IOException {
+                try {
+                    if (response.isSuccessful()) {
+
+                        String responseStr = response.body().string();
+
+                        if (responseStr.isEmpty()) {
+                            callback.onFailure(new NetworkErrorException("Empty response"));
+                        }
+                        int actualSize = currentList.size();
+                        ArrayList<Media> responseItems = getResponseFormattedList(responseStr, currentList);
+                        callback.onSuccess(filters, responseItems, responseItems.size() > actualSize);
+                        return;
+                    }
+                } catch (Exception e) {
+                    callback.onFailure(e);
+                }
+                callback.onFailure(new NetworkErrorException("Couldn't connect to API"));
+            }
+        });
+    }
+
+    public Call getDetail(ArrayList<Media> currentList, Integer index, final Callback callback) {
+        Request.Builder requestBuilder = new Request.Builder();
+        String url = apiUrls[currentApi] + itemDetailsPath + currentList.get(index).videoId;
+        requestBuilder.url(url);
+        requestBuilder.tag(MEDIA_CALL);
+
+        Timber.d(this.getClass().getSimpleName(), "Making request to: " + url);
+
+        return enqueue(requestBuilder.build(), new com.squareup.okhttp.Callback() {
+            @Override
+            public void onFailure(Request request, IOException e) {
+                callback.onFailure(e);
+            }
+
+            @Override
+            public void onResponse(Response response) throws IOException {
+                try {
+                    if (response.isSuccessful()) {
+
+                        String responseStr = response.body().string();
+
+                        if (responseStr.isEmpty()) {
+                            callback.onFailure(new NetworkErrorException("Empty response"));
+                        }
+
+                        ArrayList<Media> formattedData = getResponseDetailsFormattedList(responseStr);
+                        if (formattedData.size() > 0) {
+                            callback.onSuccess(null, formattedData, true);
+                            return;
+                        }
+                        callback.onFailure(new IllegalStateException("Empty list"));
+                        return;
+                    }
+                } catch (Exception e) {
+                    callback.onFailure(e);
+                }
+                callback.onFailure(new NetworkErrorException("Couldn't connect to API"));
+            }
+        });
+    }
+
+    public int getLoadingMessage() {
+        return R.string.loading;
+    }
+
+    public ArrayList<Media> getResponseFormattedList(String responseStr, ArrayList<Media> currentList) throws IOException {
+        return new ArrayList<>();
+    }
+
+    public ArrayList<Media> getResponseDetailsFormattedList(String responseStr) throws IOException {
+        return new ArrayList<>();
+    }
+
+    public List<NavInfo> getNavigation() {
+        return new ArrayList<>();
+    }
 
     public int getDefaultNavigationIndex() {
         return 1;
@@ -147,7 +341,7 @@ public abstract class MediaProvider extends BaseProvider implements Parcelable {
         private Filters.Order mDefOrder;
         private String mLabel;
 
-        public NavInfo(int id, Filters.Sort sort, Filters.Order defOrder, String label, @Nullable @DrawableRes Integer icon) {
+        NavInfo(int id, Filters.Sort sort, Filters.Order defOrder, String label, @Nullable @DrawableRes int icon) {
             mId = id;
             mSort = sort;
             mDefOrder = defOrder;
